funcionalidad de editar implementada - muro
