+++ src/views/login.js
@@ -66,7 +66,7 @@ export const login = () => {
     const passwordLogin = inputPass.value;
     singUser(emailLogin, passwordLogin)
       .then((userCredential) => { // Si el usuario se acredita, será dirigido al muro
-        onNavigate('/profile');
+        onNavigate('/wall');
       })
       .catch((error) => { // si hubo un error en el registro, retorna según el caso
         const errorCode = error.code;
@@ -93,7 +93,7 @@ export const login = () => {
         const token = credential.accessToken;
         // The signed-in user info.
         const user = result.user;
-        onNavigate('/profile');
+        onNavigate('/wall');
       }).catch((error) => {
         // Handle Errors here.
         const errorCode = error.code;
